docs: much needed TSDoc cleanup and remove framework notices from public exports
