+++ packages/seedcord/src/bot/gates/catalog/access.ts
@@ -1,49 +1,17 @@
+import { NotInDm, NotInGuild, NotOwner } from '@bot/notices';
+
 import { defineGate } from '../Gate';
-import { GateNotice } from './GateNotice';
 import { pickNotice } from './options';
 
 import type { GateNoticeOptions } from './options';
 import type { Gate, GateContextBase } from '../Gate';
 
 /**
- * Refusal shown when the caller is not a configured bot owner.
- *
- * @param message - Text shown in the refusal, defaulting to a bot-owner-only line.
- */
-export class NotOwner extends GateNotice {
-    public constructor(message = 'Only the bot owner can use this.') {
-        super(message);
-    }
-}
-
-/**
- * Refusal shown when a guild-only command runs in a DM.
- *
- * @param message - Text shown in the refusal, defaulting to a server-only line.
- */
-export class NotInGuild extends GateNotice {
-    public constructor(message = 'This can only be used in a server.') {
-        super(message);
-    }
-}
-
-/**
- * Refusal shown when a DM-only command runs in a guild.
- *
- * @param message - Text shown in the refusal, defaulting to a direct-message-only line.
- */
-export class NotInDm extends GateNotice {
-    public constructor(message = 'This can only be used in a direct message.') {
-        super(message);
-    }
-}
-
-/**
- * Passes only for a user id listed in `config.ownerIds`, else refuses with an ephemeral {@link NotOwner}.
+ * Passes only for a user id listed in `config.ownerIds`, else refuses.
  *
  * Agnostic, so it attaches to interaction and event handlers alike. With no `ownerIds` configured it refuses every caller. Pass {@link GateNoticeOptions} to reword or replace the refusal.
  *
- * @param options - Reword or replace the refusal. Omit to throw the default {@link NotOwner}.
+ * @param options - Reword the default refusal with `message`, or replace it with `notice`.
  *
  * @see {@link Gated}
  *
@@ -73,11 +41,11 @@ export function OwnerOnly(options?: GateNoticeOptions): Gate<GateContextBase, 'O
 }
 
 /**
- * Passes inside a guild, else refuses with {@link NotInGuild}.
+ * Passes inside a guild, else refuses.
  *
  * Agnostic, so it attaches to any handler kind. Often paired with {@link RequirePermissions} in an {@link and}.
  *
- * @param options - Reword or replace the refusal. Omit to throw the default {@link NotInGuild}.
+ * @param options - Reword the default refusal with `message`, or replace it with `notice`.
  *
  * @see {@link Gated}
  *
@@ -102,11 +70,11 @@ export function GuildOnly(options?: GateNoticeOptions): Gate<GateContextBase, 'G
 }
 
 /**
- * Passes in a direct message, else refuses with {@link NotInDm}.
+ * Passes in a direct message, else refuses.
  *
  * Agnostic, so it attaches to any handler kind. The inverse of {@link GuildOnly}, so combining the two in an {@link and} can never pass.
  *
- * @param options - Reword or replace the refusal. Omit to throw the default {@link NotInDm}.
+ * @param options - Reword the default refusal with `message`, or replace it with `notice`.
  *
  * @see {@link Gated}
  *
