+++ packages/seedcord/src/bot/gates/catalog/index.ts
@@ -1,10 +1,9 @@
-export { DmOnly, GuildOnly, NotInDm, NotInGuild, NotOwner, OwnerOnly } from './access';
-export { Cooldown, OnCooldown } from './Cooldown';
-export { GateNotice } from './GateNotice';
+export { DmOnly, GuildOnly, OwnerOnly } from './access';
+export { Cooldown } from './Cooldown';
 export { IgnoreBots } from './IgnoreBots';
-export { Nsfw, NotNsfw } from './Nsfw';
-export { MissingRole, RequireBotPermissions, RequirePermissions, RequireRole } from './permissions';
+export { Nsfw } from './Nsfw';
+export { RequireBotPermissions, RequirePermissions, RequireRole } from './permissions';
 
 export type { CooldownOptions } from './Cooldown';
 export type { GateNoticeOptions } from './options';
-export type { RequirePermissionsOptions } from './permissions';
+export type { RequirePermissionsOptions, RequireRoleOptions } from './permissions';
