+++ packages/seedcord/src/bot/utilities/permissions/checkBotPermissions.ts
@@ -1,20 +1,21 @@
 import { Logger } from '@seedcord/services';
 import { Guild } from 'discord.js';
 
+import { MissingPermissions } from '@bot/notices';
+
 import { checkPermissions, PermissionNames } from './checkPermissions';
-import { MissingPermissions } from './notices';
 
 import type { BotPermissionScope, PermissionErrorCtors } from './checkPermissions';
 import type { TextChannel } from 'discord.js';
 
 /**
- * Checks if the bot has required permissions in a {@link Guild} or {@link TextChannel}.
+ * Checks if the bot has required permissions in a {@link Guild} or {@link TextChannel}. Refuses when a
+ * required permission is missing.
  *
  * @param target - Guild or text channel to check in
  * @param scope - Permission bits to validate
  * @param inverse - Whether to check for absence of the given permissions
  * @param errors - Optional custom error constructors
- * @throws A {@link MissingPermissions} error when required permissions are missing or when bot member is unavailable/uncached
  *
  * @example
  * ```ts
