+++ packages/seedcord/src/bot/utilities/users/fetchUser.ts
@@ -1,49 +1,25 @@
-import { Notice } from '@seedcord/kit';
-import { NoticeCard } from '@seedcord/kit/internal';
 import { DiscordAPIError, RESTJSONErrorCodes } from 'discord.js';
 
-import type { ReplyResponse } from '@seedcord/types';
+import { UserNotFound } from '@bot/notices';
+
+import type { Notice } from '@seedcord/kit';
 import type { Client, User } from 'discord.js';
-
-/**
- * Error thrown when a requested user cannot be found.
- */
-class UserNotFound extends Notice {
-    constructor(public readonly userArg: string) {
-        super(`User not found: ${userArg}`);
-    }
-
-    render(): ReplyResponse {
-        const card = new NoticeCard(
-            `User probably doesn't exist or was deleted.\n` +
-                `**User Argument:** \`${this.userArg}\`\n` +
-                `Please check the user ID and try again. Only pass valid user IDs as the argument.`,
-            'User Not Found'
-        );
-        return { components: [card.component] };
-    }
-}
 
 /**
  * Options for {@link fetchUser}.
  */
 export interface FetchUserOptions {
-    /**
-     * Notice to throw when the user does not exist.
-     *
-     * @defaultValue {@link UserNotFound}
-     */
+    /** Notice shown when the user does not exist, defaulting to the standard user-not-found notice. */
     throwAs?: new (userArg: string) => Notice;
 }
 
 /**
- * Fetches a Discord user by ID with error handling.
+ * Fetches a Discord user by ID. Refuses when the user doesn't exist, with `options.throwAs` or the default.
  *
  * @param client - The Discord client instance
  * @param userId - The Discord user ID
- * @param options - Optional overrides, including the {@link FetchUserOptions.throwAs} denial
+ * @param options - Optional overrides, including the {@link FetchUserOptions.throwAs} notice
  * @returns Promise resolving to the user
- * @throws The {@link FetchUserOptions.throwAs} denial when the user doesn't exist
  */
 export async function fetchUser(client: Client, userId: string, options?: FetchUserOptions): Promise<User> {
     const Throw = options?.throwAs ?? UserNotFound;
